third_party/WebKit/LayoutTests/http/tests/websocket/useragent-in-openinghandshake.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/get-request-header.js"></script>
<script>
description("WebSocket test to check if opening handshake contains User-Agent header");

window.jsTestIsAsync = true;

var user_agent;

connectAndGetRequestHeader('user-agent').then(function(value)
{
    user_agent = value;
    shouldBe('user_agent', 'navigator.userAgent');
    finishJSTest();
}, finishAsFailed);
</script>
