third_party/blink/web_tests/http/tests/websocket/useragent-in-openinghandshake-expected.txt

WebSocket test to check if opening handshake contains User-Agent header

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS user_agent is navigator.userAgent
PASS successfullyParsed is true

TEST COMPLETE
